Moved out ProjectInfo directory/component from common
