--- src/common/components/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectInfo } from './ProjectInfo'
